♻️ refactor : 선호 지역 관리 기능 Controller/Service 분리
[분리 내용]
- UserPreferredRegionController 신규 생성
  - GET /preferred-regions/view: 선호 지역 조회
  - PUT /preferred-regions/reselect: 선호 지역 변경

- UserPreferredRegionService 신규 생성
  - getUserPreferredRegions(): 선호 지역 조회 로직
  - updateUserPreferredRegions(): 선호 지역 변경 로직
  - setPreferredRegionsForOnboarding(): 온보딩용 지역 설정

[리팩토링 효과]
- 단일 책임 원칙(SRP) 준수
  - OnboardingService: 온보딩 프로세스 진행에만 집중
  - UserPreferredRegionService: 선호 지역 관리에만 집중
- 응집도 향상 및 결합도 감소
- 마이페이지 기능 독립적 관리 가능

[변경 사항]
- OnboardingController: 선호 지역 관련 API 제거
- OnboardingService:
  - 선호 지역 관련 메소드 제거
  - UserPreferredRegionService 주입 및 위임
- HTTP 테스트 파일 업데이트 (onboarding.http)

[API 엔드포인트 변경]
- 기존: /onboarding/regions → 신규: /preferred-regions
- 온보딩 과정 중 지역 선택: /onboarding/region (유지)
- 마이페이지 지역 관리: /preferred-regions (신규)

Refs: #174, #156, #63, #55
